Move enabeld to be a StateNode feature.
Moved enabled to be a state node feature
instead of a NodeFeature.

Added feature check to Component default
onEnabledStateChange that we can actually
set the attrubute.
