+++ flow-server/src/main/java/com/vaadin/flow/dom/impl/AbstractTextElementStateProvider.java
@@ -208,14 +208,4 @@ public abstract class AbstractTextElementStateProvider
         throw new UnsupportedOperationException();
     }
 
-    @Override
-    public void setEnabled(StateNode node, boolean enabled) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public boolean isEnabled(StateNode node) {
-        throw new UnsupportedOperationException();
-    }
-
 }
